+++ src/TrelloDotNet/Model/Options/CardFields.cs
@@ -1,4 +1,5 @@
-﻿using System.Linq;
+﻿using System.Collections.Generic;
+using System.Linq;
 using TrelloDotNet.Control;
 
 namespace TrelloDotNet.Model.Options
@@ -12,6 +13,33 @@ namespace TrelloDotNet.Model.Options
         /// Fields to include
         /// </summary>
         internal string[] Fields { get; set; }
+
+        /// <summary>
+        /// The default Fields the API return if not specified
+        /// </summary>
+        internal static CardFields DefaultFields => new CardFields(
+            CardFieldsType.DueComplete,
+            CardFieldsType.Closed,
+            CardFieldsType.Description,
+            CardFieldsType.BoardId,
+            CardFieldsType.ChecklistIds,
+            CardFieldsType.ListId,
+            CardFieldsType.MemberIds,
+            CardFieldsType.LabelIds,
+            CardFieldsType.Labels,
+            CardFieldsType.IdShort,
+            CardFieldsType.MembersVotedIds,
+            CardFieldsType.AttachmentCover,
+            CardFieldsType.Name,
+            CardFieldsType.ShortUrl,
+            CardFieldsType.Url,
+            CardFieldsType.Start,
+            CardFieldsType.Cover,
+            CardFieldsType.Due,
+            CardFieldsType.IsTemplate,
+            CardFieldsType.Position,
+            CardFieldsType.Subscribed,
+            CardFieldsType.LastActivity);
 
         /// <summary>
         /// All Fields
@@ -34,6 +62,18 @@ namespace TrelloDotNet.Model.Options
         public CardFields(params CardFieldsType[] fields)
         {
             Fields = fields.Select(x => x.GetJsonPropertyName()).ToArray();
+        }
+
+        internal void AddIfMissing(CardFieldsType field)
+        {
+            var propertyName = field.GetJsonPropertyName();
+            if (!Fields.Contains(propertyName))
+            {
+                Fields = Fields.Union(new List<string>
+                {
+                    propertyName
+                }).ToArray();
+            }
         }
     }
 }